+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/FirefoxFocus.kt
@@ -48,6 +48,5 @@ class FirefoxFocus(private val apiConsumer: ApiConsumer) : BaseAppDetail() {
     }
 
     companion object {
-        const val INSTALLED_VERSION_KEY = "device_app_register_FIREFOX_FOCUS_version_name"
     }
 }