+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 2021-XX-XX Version 72.1.0 (61)
+ * Add support for Bromite https://github.com/bromite/bromite (Tobiwan/ffupdater#59 Tobi823/ffupdater#22)
  * Improve Brazilian Portuguese translation (thanks mezysinc; Tobiwan/ffupdater#58)
  * Improve Russian translation (thanks DeenHyper74; Tobiwan/ffupdater#56)
- * Make download status translatable (thanks DeenHyper74; Tobiwan/ffupdater#57)
- * Ask for confirmation when the user wants to update an app but the latest app version is already installed (thanks DeenHyper74; Tobiwan/ffupdater#60)
+ * Make download status translatable (Tobiwan/ffupdater#57)
+ * Ask for confirmation when the user wants to update an app but the latest app version is already installed (Tobiwan/ffupdater#60)
 
 # 2021-03-14 Version 72.0.0 (60)
  * !!! Remove support for Firefox Light because updates are no longer signed and therefore pose a security risk (thanks opened and mega-stoffel)